+++ CHANGELOG.md
@@ -17,6 +17,11 @@ Changelog structure reference:
 
 ## [Unreleased]
 
+### Added
+
+- Project
+  - PTY emulation for non-debug runs. Fixes colored output in Ziglings.
+
 ## [15.0.3]
 
 ### Fixed